+++ infrastructure/modules/armonik/admin-gui.tf
@@ -1,5 +1,6 @@
 # Control plane deployment
 resource "kubernetes_deployment" "admin_gui" {
+  count = var.admin_gui != null ? 1 : 0
   metadata {
     name      = "admin-gui"
     namespace = var.namespace
@@ -70,17 +71,17 @@ resource "kubernetes_deployment" "admin_gui" {
             value = local.control_plane_url
           }
           dynamic "env" {
-            for_each = (local.grafana_url != "" ? [1] : [])
+            for_each = (data.kubernetes_secret.grafana.data.enabled ? [1] : [])
             content {
               name  = "Grafana__Endpoint"
-              value = local.grafana_url
+              value = data.kubernetes_secret.grafana.data.url
             }
           }
           dynamic "env" {
-            for_each = (local.seq_web_url != "" ? [1] : [])
+            for_each = (data.kubernetes_secret.seq.data.enabled ? [1] : [])
             content {
               name  = "Seq__Endpoint"
-              value = local.seq_web_url
+              value = data.kubernetes_secret.seq.data.web_url
             }
           }
           dynamic "env" {
@@ -128,17 +129,17 @@ resource "kubernetes_deployment" "admin_gui" {
             value = local.control_plane_url
           }
           dynamic "env" {
-            for_each = (local.grafana_url != "" ? [1] : [])
+            for_each = (data.kubernetes_secret.grafana.data.enabled != "" ? [1] : [])
             content {
               name  = "Grafana__Endpoint"
-              value = local.grafana_url
+              value = data.kubernetes_secret.grafana.data.url
             }
           }
           dynamic "env" {
-            for_each = (local.seq_web_url != "" ? [1] : [])
+            for_each = (data.kubernetes_secret.seq.data.enabled ? [1] : [])
             content {
               name  = "Seq__Endpoint"
-              value = local.seq_web_url
+              value = data.kubernetes_secret.seq.data.web_url
             }
           }
           dynamic "env" {
@@ -181,34 +182,34 @@ resource "kubernetes_deployment" "admin_gui" {
 
 # Admin GUI service
 resource "kubernetes_service" "admin_gui" {
+  count = length(kubernetes_deployment.admin_gui)
   metadata {
-    name      = kubernetes_deployment.admin_gui.metadata.0.name
-    namespace = kubernetes_deployment.admin_gui.metadata.0.namespace
+    name      = kubernetes_deployment.admin_gui[0].metadata.0.name
+    namespace = kubernetes_deployment.admin_gui[0].metadata.0.namespace
     labels = {
-      app     = kubernetes_deployment.admin_gui.metadata.0.labels.app
-      service = kubernetes_deployment.admin_gui.metadata.0.labels.service
+      app     = kubernetes_deployment.admin_gui[0].metadata.0.labels.app
+      service = kubernetes_deployment.admin_gui[0].metadata.0.labels.service
     }
   }
   spec {
     type = var.admin_gui.service_type
     selector = {
-      app     = kubernetes_deployment.admin_gui.metadata.0.labels.app
-      service = kubernetes_deployment.admin_gui.metadata.0.labels.service
+      app     = kubernetes_deployment.admin_gui[0].metadata.0.labels.app
+      service = kubernetes_deployment.admin_gui[0].metadata.0.labels.service
     }
     port {
-      name        = kubernetes_deployment.admin_gui.spec.0.template.0.spec.0.container.0.port.0.name
+      name        = kubernetes_deployment.admin_gui[0].spec.0.template.0.spec.0.container.0.port.0.name
       port        = var.admin_gui.api.port
-      target_port = kubernetes_deployment.admin_gui.spec.0.template.0.spec.0.container.0.port.0.container_port
+      target_port = kubernetes_deployment.admin_gui[0].spec.0.template.0.spec.0.container.0.port.0.container_port
       protocol    = "TCP"
     }
     port {
-      name        = kubernetes_deployment.admin_gui.spec.0.template.0.spec.0.container.1.port.0.name
+      name        = kubernetes_deployment.admin_gui[0].spec.0.template.0.spec.0.container.1.port.0.name
       port        = var.admin_gui.app.port
-      target_port = kubernetes_deployment.admin_gui.spec.0.template.0.spec.0.container.1.port.0.container_port
+      target_port = kubernetes_deployment.admin_gui[0].spec.0.template.0.spec.0.container.1.port.0.container_port
       protocol    = "TCP"
     }
   }
-
   timeouts {
     create = "2m"
   }