+++ infrastructure/modules/onpremise-storage/minio/variables.tf
@@ -5,16 +5,14 @@ variable "namespace" {
 }
 
 # Parameters for minio
-variable "minioconfig" {
+variable "minio" {
   description = "Parameters of S3 payload storage"
   type = object({
-    image         = string
-    tag           = string
-    host          = string
-    port          = string
-    login         = string
-    password      = string
-    bucket_name   = string
-    node_selector = any
+    image              = string
+    tag                = string
+    image_pull_secrets = string
+    host               = string
+    bucket_name        = string
+    node_selector      = any
   })
 }