+++ infrastructure/quick-deploy/aws/storage/s3-iam-object-storage.tf
@@ -1,5 +1,6 @@
 # Decrypt objects in S3
 data "aws_iam_policy_document" "decrypt_s3_storage_object" {
+  count = length(module.s3_os) > 0 ? 1 : 0
   statement {
     sid = "KMSAccess"
     actions = [
@@ -17,19 +18,22 @@ data "aws_iam_policy_document" "decrypt_s3_storage_object" {
 }
 
 resource "aws_iam_policy" "decrypt_s3_storage_object" {
+  count       = length(module.s3_os) > 0 ? 1 : 0
   name_prefix = local.iam_s3_decrypt_s3_storage_object_policy_name
   description = "Policy for alowing decryption of encrypted object in S3 ${var.eks.cluster_id}"
-  policy      = data.aws_iam_policy_document.decrypt_s3_storage_object.json
+  policy      = data.aws_iam_policy_document.decrypt_s3_storage_object[0].json
   tags        = local.tags
 }
 
 resource "aws_iam_role_policy_attachment" "decrypt_s3_storage_object" {
-  policy_arn = aws_iam_policy.decrypt_s3_storage_object.arn
+  count      = length(module.s3_os) > 0 ? 1 : 0
+  policy_arn = aws_iam_policy.decrypt_s3_storage_object[0].arn
   role       = var.eks.worker_iam_role_name
 }
 
 # Read objects in S3
 data "aws_iam_policy_document" "fullaccess_s3_storage_object" {
+  count = length(module.s3_os) > 0 ? 1 : 0
   statement {
     sid = "FullAccessFromS3"
     actions = [
@@ -42,19 +46,21 @@ data "aws_iam_policy_document" "fullaccess_s3_storage_object" {
     ]
     effect = "Allow"
     resources = [
-      "${module.s3_os.arn}/*"
+      "${module.s3_os[0].arn}/*"
     ]
   }
 }
 
 resource "aws_iam_policy" "fullaccess_s3_storage_object" {
+  count       = length(module.s3_os) > 0 ? 1 : 0
   name_prefix = "s3-fullAccess-${var.eks.cluster_id}"
   description = "Policy for allowing read/write/delete object in S3 ${var.eks.cluster_id}"
-  policy      = data.aws_iam_policy_document.fullaccess_s3_storage_object.json
+  policy      = data.aws_iam_policy_document.fullaccess_s3_storage_object[0].json
   tags        = local.tags
 }
 
 resource "aws_iam_role_policy_attachment" "fullaccess_s3_storage_object_attachment" {
-  policy_arn = aws_iam_policy.fullaccess_s3_storage_object.arn
+  count      = length(module.s3_os) > 0 ? 1 : 0
+  policy_arn = aws_iam_policy.fullaccess_s3_storage_object[0].arn
   role       = var.eks.worker_iam_role_name
 }