+++ infrastructure/quick-deploy/localhost/storage/variables.tf
@@ -61,28 +61,19 @@ variable "redis" {
     image_pull_secrets = string
     max_memory         = string
   })
+  default = null
 }
 
 # Parameters for minio
-variable "minioconfig" {
+variable "minio" {
   description = "Parameters of minio"
   type = object({
-    image         = string
-    tag           = string
-    host          = string
-    port          = string
-    login         = string
-    password      = string
-    bucket_name   = string
-    node_selector = any
+    image              = string
+    tag                = string
+    image_pull_secrets = string
+    host               = string
+    default_bucket     = string
+    node_selector      = any
   })
-}
-
-variable "object_storages_to_be_deployed" {
-  description = "The list of storage objects to be deployed"
-  type        = list(string)
-  validation {
-    condition     = alltrue([for value in var.object_storages_to_be_deployed : contains(["mongodb", "redis", "s3", "localstorage"], lower(value))])
-    error_message = "Valid values for object_storages_to_be_deployed are \"MongoDB\" | \"Redis\" | \"S3\" | \"LocalStorage\"."
-  }
+  default = null
 }